requirements.txt
aistudio-sdk==0.3.8
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
bce-python-sdk==0.9.59
certifi==2026.1.4
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
distro==1.9.0
executing==2.2.1
fastapi==0.109.0
filelock==3.20.3
fire==0.5.0
google-cloud-storage==2.18.2
fonttools==4.61.1
fsspec==2026.1.0
future==1.0.0
google-auth==2.47.0
google-genai==1.59.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.2
idna==3.11
imagesize==1.4.1
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
langchain-core==1.2.7
langchain-text-splitters==1.1.0
langsmith==0.6.4
MarkupSafe==2.1.5
matplotlib==3.10.8
matplotlib-inline==0.2.1
modelscope==1.34.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.11.0.86
opencv-python-headless==4.13.0.90
opt-einsum==3.3.0
orjson==3.11.5
packaging==25.0
paddleocr==3.3.0
paddlepaddle==3.2.1
paddlex==3.3.13
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
Pillow==9.5.0
platformdirs==4.5.1
prettytable==3.17.0
prompt_toolkit==3.0.52
protobuf==6.33.4
psutil==7.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pyclipper==1.4.0
pycryptodome==3.23.0
pydantic==2.9.2
pydantic_core==2.23.4
pydantic-settings==2.6.1
Pygments==2.19.2
pyparsing==3.3.2
pypdfium2==5.3.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
ruamel.yaml==0.19.1
safetensors==0.7.0
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
simple-lama-inpainting==0.1.2
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.35.1
sympy==1.14.0
tenacity==9.1.2
termcolor==3.3.0
torch==2.10.0+cpu
torchvision==0.25.0+cpu
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.27.0
wcwidth==0.2.14
websockets==15.0.1
zstandard==0.25.0
